Mrak & Dron...

+++ projects/commonlibs/zocromas_mas_multiconfig/src/inc/muc_option_extended.h
@@ -3,6 +3,6 @@
 
 # include "muc_option_types.h"                                       /* muc_longval_extended_vtable_t */
 
-int muc_longindex_extended_count( muc_longval_extended_vtable_t ** xvtables );
+size_t muc_longindex_extended_count( muc_longval_extended_vtable_t ** xvtables );
 
 #endif
